Avoid 'jq' tool utilization during 'pre-config' stage
Prior to this PR, we were relying on 'jq' tool during the pre-configuration stage (i.e. sysbox.config) to identify the setup on which to install Sysbox (i.e. 'userns' vs 'shiftfs', etc). This was causing
installation issues in scenarios where 'jq' was not present ahead of time.

Unfortunately, dependency issues at this early stage cannot be dealt with by the typical 'Depend' or 'Pre-Depend' clause in the debian's control file, as those are only useful to satisfy dependencies required
by the pre-installation (sysbox.preinst) and post-installation (sysbox.postinst) stages. As per debian-policy doc:

"The config script might be run before the preinst script and before the package is unpacked or any of its dependencies or pre-dependencies are satisfied. Therefore it must work using only the tools present in essential packages."

The solution in this case is to leave out 'jq' from the pre-config phase and rely on 'grep' instructions instead.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>