Expand installer to account for 'configfs' kernel module dependency
Prior to this PR, we had always assumed that 'configfs' was loaded by default, but in reality that's not always the case outside the Ubuntu camp.

Now, 'configfs' is not strictly needed by Sysbox per-se, we do it just to satisfy certain applications that may be interested in reading/writing from/to the typical configfs path
(i.e. `/sys/kernel/config`). Now, technically speaking, we shouldn't need configfs module at all, coz all we do is to expose a dummy dir (i.e. 'tmpfs' bind-mount) in the configfs
path. Unfortunately, the kernel doesn't allow anyone (not even root) to create a dir inside the sysfs hierarchy, so that's why we end up setting configfs module as a dependency.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>