Refactor and expand Sysbox-packager to generate both CE and EE packages
As part of this effort, i initially spent some time trying to expand my understanding of debian-policy guidelines to allow us to generate different packages by making use of the same set of installation scripts (i.e. w/o duplicating most of the scripts), but that
proved to be a more complex approach that would require more study and won't necessarily pay off in the near term. I ended up duplicating the installation scripts (one set for CE and another one for EE), which will eventually allow us to customize the installation
process for each Sysbox product.

As a consequence, now we have two different debian packages for Sysbox: 'sysbox-ce' and 'sysbox-ee'.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>