Workaround old Makefile issue
I'm ashamed to admit that I haven't found a solution for this one since the first time i reproduced it more than a year ago :-(

Problem seems to be related to the recursive layout in which Makefiles are placed in this module. Problem is reproduced by executing a target in the top-most Makefile, that in turn has a dependency on an inner folder. In the output displayed below, i've commented out the main Makefile instruction to focus on the initialization and finalization of the target execution.

Notice that the actions associated to the inner target are properly executed, but error is observed when returning back to the outer Makefile.

```
rmolina@dev-vm1:~/wsp/03-30-2021/sysbox-internal/sysbox-pkgr$ make sysbox-ee-deb ubuntu-focal
make -C deb --no-print-directory clean
make: *** No rule to make target 'ubuntu-focal'.  Stop.
```

For some reason that i'm yet to understand, problem can be workaround'ed by exposing the targets in the inner folders as 'phony' targets in the outer Makefile..

Signed-off-by: Rodny Molina <rmolina@nestybox.com>