Modify sysbox package to have sysbox send logs to the systemd log.
Recent changes in sysbox-mgr and sysbox-fs enable them to log
to the systemd journal log (by virtue of logging on stderr
by default). This is advantageous as we don't have to worry
about log size or log rotation (systemd takes care of that).

In this change, we modify the systemd unit files for
sysbox-mgr and sysbox-fs such that they will no longer
log to the "/var/log/sysbox-*" files and instead logs
will now go to the systemd journal log.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>