Modified sysbox-deploy-k8s daemonset to restart CRI-O after config.
Prior to this change the daemonset was sending a SIGHUP signal to CRI-O after
adding the "sysbox-runc" config to the /etc/crio/crio.conf file.

But unfortunately CRI-O does not support "live config reload" for
the runtime config. As a result, the only choice is to restart
CRI-O, which is a more disruptive operation as it causes all
pods on the K8s node to be torn-down and restarted.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>