Add logic in crio-deploy-k8s to eliminate all the state associated with the runtime being replaced
Problem was initially observed in an AWS cluster that utilized Calico as
CNI. During problem reproduction none of the Calico components were able to
execute in those nodes where CRI-O had been installed.

We noticed that Calico was complaining about pre-existing instances of its
components running in the nodes. It turns out that the Calico pods previously
launched with the original runtime (containerd), were still running in the
affected nodes.

The solution to this problem, which is not Calico specific and applies to all
other daemonsets/regular-pods, is to ensure that the state associated with
non-active (or soon-to-be-idle) runtimes is properly eliminated as part of the
CRI-O daemonset installation (and uninstallation).

Signed-off-by: Rodny Molina <rmolina@nestybox.com>