crio-deploy-k8s: stop containerd and link it's socket to the CRI-O socket.
When installing CRI-O on a K8s worker node, the crio-deploy-k8s daemonset
needs to also stop containerd and in addition create a soft-link
from the containerd socket to the CRI-O socket.

This soft-link is important because some pods (e.g., GKE's gke-metadata-server)
assume that containerd is the runtime on the node, and talk to it via a
containerd socket mount to the pod. Without this soft-link, such pods fail when
CRI-O replaces containerd. With the soft-link, those same pods work because they
talk via the containerd socket to CRI-O.  Since both containerd and1 CRI-O follow
the CRI spec, the pods can't tell between one or the other.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>