Enhance state-cleaning logic in sysbox-pod scenarios with dockershim CRI
The main purpose here is to offer a more robust approach to the state-cleaning logic executed during "dockershim to cri-o" transitions.

In the previous approach we avoid stopping kubelet and expect that no new state arrives in the interval between the state-cleaning and the kubelet being restarted with the cri-o configuration. This approach is not ideal as it's prone to race-conditions that end up with stale entries in the control plane.

In this new logic we do stop kubelet, but before doing so, we obtain a list of the pods created with the pre-existing runtime, and then, once that kubelet is down, we eliminate the associated containers through the docker-cli interface.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>