[k8s] Re-add crio-extractor script to sysbox-ce Dockerfile
This one was removed a few days ago as, at that time, the crio-extractor script was expected to be used only in Flatcar scenarios, and these ones would be supported in Sysbox-EE only. But now that we are embedding the crio binaries within both CE and EE docker-images, this extractor script needs to be re-added to the CE image.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>