[k8s] Fix for scenario with more than one env-var holding 'container-runtime' attributes
The existing logic assumes that there will always be one single env-var holding the `--container-runtime-endpoint` attribute. But this is not the case in AKS, where an existing env-var (different from the regular KUBELET_EXTRAS) is already storing this attribute.

In consequence, when we try to revert the kubelet config we find ourselves with two container-runtime attributes (containerd and crio), and we end up picking containerd (the first one in the list), which causes a failure as 'crio' should be the one being displayed when iterating through the config-revert logic.

```
Environment=KUBELET_CONTAINERD_FLAGS="--container-runtime=remote --runtime-request-timeout=15m --container-runtime-endpoint=unix:///run/containerd/containerd.sock"
```

Signed-off-by: Rodny Molina <rmolina@nestybox.com>