Improve CRI-O / conmon config file installation.
This changes improves the way we setup the CRI-O and conmon config files, by
copying the files to the host where CRI-O is installed, rather than creating
portions of the files in the crio installation script via bash heredoc.

This changes also includes a few config files that we had previously missed.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>