[k8s] Enhance state-cleanup logic for better coverage of dockershim setups (RKE)
I'm simply merging the new clean_runtime_state_dockershim() function with its older version as I previously broke the RKE setup while trying to add support for RKE2.

In essence, we can't simply wipe out all the existing containers coz that also eliminates the kubelet container that we previously re-configured. The fix is to allow this function to operate in two modes:

   * Implicit: If no pod-list is passed by the caller then we'll eliminate all the existing containers.
   * Explicit: If a pod-list is provided, then only eliminate its associated containers (RKE case).

Signed-off-by: Rodny Molina <rmolina@nestybox.com>