[k8s] Change CentOS' sysbox-deploy underlying image to address multi-arch requirements
Prior to this change, sysbox-deploy was relying on the 'centos/systemd' image as the baseline over which to install/build all the sysbox components/dependencies. Problem with this particular image is that it only supports 'amd64' platforms.

To fix this limitation I've compiled our own 'centos/systemd' image for each of the supported hardware platforms. These two images have been combined through the utilization of a docker-manifest (same as we do with multi-arch images utilized in sysbox ITs). This manifest will be the one provided to users so that the proper docker-image is downloaded during sysbox-deploy instantiation.

The new 'centos/systemd' file has been renamed as 'ghcr.io/nestybox/centos7/systemd', and its associated Dockerfile has been placed in 'sysbox-pkgr/k8s' directory.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>