[k8s] Fix recently-introduced bug affecting kernel-comparison logic
In a recent commit we modified the routine that extracts the kernel-version running in the host to collect the entire string (e.g., 5.13.0-39-generic). This is useful to detect kernel upgrades and force sysbox-deploy reinstalation process. Unfortunately, this caused issues with the current `semver_lt/ge` routines as the length of the strings to compare is undetermined.

Instead of adapting these semver functions, we are simply truncating the kernel-version string to obtain only the kernel's major and minor numbers (e.g., 5.13).

Signed-off-by: Rodny Molina <rmolina@nestybox.com>