Change the pods-per-node limit on Sysbox-CE from 16 to 4K.
Prior to this change, Sysbox-CE was artificially limited to 16-pods-per-node
via the sysbox-deploy-k8s script. After this change, that limit is bumped
to 4K-pod-per-node.

This matches Sysbox-CE with Sysbox-EE, and it's being done to encourage
Sysbox-CE adoption in K8s clusters particularly since Sysbox-EE is no longer
offered as a standalone product as of 05/22. Decision to do this was agreed by
Cesar Talledo, Rodny Molina, Jake Levirne, and Justim Cormack at Docker.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>