Add detection for RKE2 controller service in daemonset installer
Previously, only `rke2-agent.service` was checked for, but in a single node
cluster there is only the controller node with the `rke2-server` service
running instead. This adds support for checking for `rke2-server` and handles
stopping/starting whichever service is running on the node.

See https://github.com/nestybox/sysbox/issues/692.