[k8s] Add support for incremental Sysbox upgrades
By 'incremental', I'm simply referring to the possibility of deleting the `sysbox-deploy-k8s' pod (through `kubectl delete sysbox-install`) and then re-applying it again (`kubectl apply sysbox-install`), without the need to re-install or restart cri-o.

Before this change, for Sysbox to be updated, we first needed to entirely uninstall Sysbox through the `sysbox-install` and `sysbox-uninstall` yamls, which is a tedious/slow process.

Proposed changes have been successfully tested in EKS.

Signed-off-by: Rodny Molina <rodny.molina@docker.com>