k8s: Delete preexisting sysbox pods during the upgrade and uninstallation process
During sysbox-deploy-k8s' update process, the Sysbox daemons can be restarted after the new artifacts are installed. However, pre-existing sysbox pods are not restarted by this process, and in consequence, they are left in an inconsistent state in which they cannot execute any of the syscalls trapped/emulated by Sysbox. This is a day-one issue, but it is much more visible now that we support Sysbox upgrades.

Signed-off-by: Rodny Molina <rodny.molina@docker.com>