k8s: Enhance sysbox-deploy-k8s 'update' logic to restart pods when Sysbox is installed from scratch
Our goal here is to ensure that, during sysbox-deploy-k8s execution, preexisting sysbox pods are only restarted when required. For example, we don't want to restart these pods if the code of daemon-set itself is changed. We would like to do so only when either Sysbox binaries, its dependencies, or its configuration has been changed.

Before this commit, the above requirement was not being met in scenarios where sysbox-deploy-k8s was being installed from scratch (we were not restarting pods at all). The reason for that is that we were relying on the 'do_sysbox_install' variable to detect new Sysbox install ions. However, this global var is not good enough since its value is "lost" when all the pods in the node are restarted during crio installation. That's why we need to introduce another variable that captures the ongoing/partial state of the installation. This new variable can be taken into account when trying to determine if sysbox pods need to be restarted or not.

Signed-off-by: Rodny Molina <rodny.molina@docker.com>