Add support for k8s v1.31 and deprecate k8s v1.27.
Add support for Kubernetes v1.31 in sysbox-deploy-k8s. Also, remove support for
K8s v1.27.

Note that in order to support K8s v1.31, we have to use a patched CRI-O "pinns"
binary (a helper to the crio binary) to overcome a problem where the K8s kubelet
is setting a sysctl (e.g., /proc/sys/net/core/somaxconn) that is not present in
K8s nodes with older kernels (e.g., 5.15) when unsharing the user-namespace. See
this commit in the Nestybox CRI-O repository for more info:
https://github.com/nestybox/cri-o/commit/25d4daec8047ee0807a4ed0ac66ed917b89afc85

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>